deepfake-therapist/backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0

# Realtime
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.2.0

# Local LLM + utilities
httpx>=0.27.0

# Speech-to-text (local Whisper)
faster-whisper>=1.0.0

# Optional local TTS Python fallback (if using Coqui engine)
# TTS>=0.22.0
